Revert "Asset deprecation (#639)"
This reverts commit db7a8c62451171244b66dfbc72240ca6e408a073.
