+++ VERSION
@@ -1,1 +1,1 @@
-0.11.2
+0.12.0-next.revert-decodable-asset
