+++ android/player/src/main/java/com/intuit/playerui/android/AndroidPlayer.kt
@@ -8,7 +8,7 @@ import com.intuit.hooks.HookContext
 import com.intuit.hooks.SyncBailHook
 import com.intuit.hooks.SyncHook
 import com.intuit.hooks.SyncWaterfallHook
-import com.intuit.playerui.android.asset.DecodableAsset.AsyncHydrationTrackerPlugin
+import com.intuit.playerui.android.asset.SuspendableAsset.AsyncHydrationTrackerPlugin
 import com.intuit.playerui.android.asset.RenderableAsset
 import com.intuit.playerui.android.extensions.Styles
 import com.intuit.playerui.android.extensions.overlayStyles
