+++ android/player/src/test/java/com/intuit/playerui/android/renderer/BrokenAssetTest.kt
@@ -12,10 +12,8 @@ import com.intuit.playerui.android.utils.TestAssetsPlugin
 import com.intuit.playerui.core.player.PlayerException
 import com.intuit.playerui.core.player.state.ErrorState
 import com.intuit.playerui.utils.start
-import com.intuit.playerui.utils.test.runBlockingTest
 import io.mockk.impl.annotations.MockK
 import io.mockk.junit5.MockKExtension
-import kotlinx.coroutines.runBlocking
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.BeforeEach
@@ -46,23 +44,21 @@ internal class BrokenAssetTest {
     @Test
     fun `invalidate view should fail on first render`() {
         assertThrows<StaleViewException> {
-            runBlocking {
-                BrokenAsset(baseContext.copy(asset = runtime.asset(shouldFail = true))).render(mockContext)
-            }
+            BrokenAsset(baseContext.copy(asset = runtime.asset(shouldFail = true))).render(mockContext)
         }
     }
 
     @Test
-    fun `invalidate view should handle gracefully in a rehydrate (if asset renders properly the second time)`(): Unit = runBlockingTest {
+    fun `invalidate view should handle gracefully in a rehydrate (if asset renders properly the second time)`() {
         assertTrue(BrokenAsset(baseContext.copy(asset = runtime.asset(layout = BrokenAsset.Layout.Frame))).render(mockContext) is FrameLayout)
         assertTrue(BrokenAsset(baseContext.copy(asset = runtime.asset(layout = BrokenAsset.Layout.Linear))).render(mockContext) is LinearLayout)
     }
 
     @Test
-    fun `manual rehydration should fail the player on invalidate view`(): Unit = runBlockingTest {
+    fun `manual rehydration should fail the player on invalidate view`() {
         BrokenAsset(baseContext.copy(asset = runtime.asset(layout = BrokenAsset.Layout.Frame))).apply {
             assertTrue(render(mockContext) is FrameLayout)
-            getData().layout = BrokenAsset.Layout.Linear
+            data.layout = BrokenAsset.Layout.Linear
             rehydrate()
         }
         assertTrue(player.state is ErrorState)
