+++ android/testutils/src/main/java/com/intuit/playerui/android/testutils/asset/Assertions.kt
@@ -1,8 +1,8 @@
 package com.intuit.playerui.android.testutils.asset
 
 import android.view.View
-import com.intuit.playerui.android.asset.DecodableAsset
 import com.intuit.playerui.android.asset.RenderableAsset
+import com.intuit.playerui.android.asset.SuspendableAsset
 import com.intuit.playerui.core.player.Player
 import com.intuit.playerui.core.player.state.PlayerFlowState
 import kotlinx.coroutines.delay
@@ -24,7 +24,7 @@ public inline fun <reified T : RenderableAsset> Any?.shouldBeAsset(
 
 @OptIn(ExperimentalContracts::class)
 public inline fun <reified T : View> Any?.shouldBeView(assertions: T.() -> Unit = {}): T {
-    val view = if (T::class != DecodableAsset.AsyncViewStub::class && this is DecodableAsset.AsyncViewStub) {
+    val view = if (T::class != SuspendableAsset.AsyncViewStub::class && this is SuspendableAsset.AsyncViewStub) {
         runBlocking {
             awaitView()
         }
