+++ plugins/reference-assets/android/src/androidTest/java/com/intuit/playerui/android/reference/assets/info/InfoTest.kt
@@ -10,7 +10,7 @@ import com.intuit.playerui.android.testutils.asset.shouldBeAsset
 import com.intuit.playerui.android.testutils.asset.shouldBeAtState
 import com.intuit.playerui.android.testutils.asset.shouldBeView
 import com.intuit.playerui.core.player.state.InProgressState
-import kotlinx.coroutines.runBlocking
+import kotlinx.coroutines.test.runTest
 import org.junit.Assert.assertEquals
 import org.junit.Test
 
@@ -20,7 +20,7 @@ class InfoTest : AssetTest("info") {
         Continue, Dismiss, Next
     }
 
-    private suspend fun verifyAndProceed(view: Int, action: PlayerAction? = null) {
+    private fun verifyAndProceed(view: Int, action: PlayerAction? = null) {
         val infoTitle = currentView?.findViewById<FrameLayout>(R.id.info_title) ?: throw AssertionError("current view is null")
         val infoFooter =
             currentView?.findViewById<FrameLayout>(R.id.info_footer) ?: throw AssertionError("current view is null")
@@ -29,19 +29,21 @@ class InfoTest : AssetTest("info") {
             assertEquals("View $view", text.toString())
         }
 
-        currentAssetTree.shouldBeAsset<Info> {
-            val infoActions = getData().actions
+        runTest {
+            currentAssetTree.shouldBeAsset<Info> {
+                val infoActions = getData().actions
 
-            action?.let {
-                val buttonOrdinal = if (action.ordinal != 1) 0 else action.ordinal
-                infoActions[buttonOrdinal].shouldBeAsset<Action> {
-                    val data = getData()
-                    data.run()
+                action?.let {
+                    val buttonOrdinal = if (action.ordinal != 1) 0 else action.ordinal
+                    infoActions[buttonOrdinal].shouldBeAsset<Action> {
+                        val data = getData()
+                        data.run()
+                    }
                 }
             }
-        }
-        infoFooter[0].shouldBeView<TextView> {
-            assertEquals("Footer Text", text.toString())
+            infoFooter[0].shouldBeView<TextView> {
+                assertEquals("Footer Text", text.toString())
+            }
         }
     }
 
@@ -49,14 +51,12 @@ class InfoTest : AssetTest("info") {
     fun infoBasic() {
         launchMock("info-modal-flow")
 
-        runBlocking {
-            verifyAndProceed(1, PlayerAction.Continue)
-            verifyAndProceed(2, PlayerAction.Dismiss)
-            verifyAndProceed(1, PlayerAction.Continue)
-            verifyAndProceed(2, PlayerAction.Next)
-            verifyAndProceed(3, PlayerAction.Next)
-            verifyAndProceed(1)
-        }
+        verifyAndProceed(1, PlayerAction.Continue)
+        verifyAndProceed(2, PlayerAction.Dismiss)
+        verifyAndProceed(1, PlayerAction.Continue)
+        verifyAndProceed(2, PlayerAction.Next)
+        verifyAndProceed(3, PlayerAction.Next)
+        verifyAndProceed(1)
 
         player.shouldBeAtState<InProgressState>()
     }
