+++ plugins/reference-assets/android/src/main/java/com/intuit/playerui/android/reference/assets/action/Action.kt
@@ -6,6 +6,7 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.material.Button
 import androidx.compose.material.ButtonDefaults
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
@@ -36,14 +37,13 @@ class Action(assetContext: AssetContext) : ComposableAsset<Action.Data>(assetCon
 
     @Composable
     override fun content(data: Data) {
+        val scope = rememberCoroutineScope()
         Button(
             onClick = {
-                hydrationScope.launch {
-                    withContext(Dispatchers.Default) {
-                        beacon("clicked", "button")
-                        player.commitPendingTransaction()
-                        data.run()
-                    }
+                beacon("clicked", "button")
+                player.commitPendingTransaction()
+                scope.launch {
+                    data.run()
                 }
             },
             modifier = Modifier.fillMaxWidth().testTag("action"),
